qified - chore: moving memory message provider to new structure
